fix: problemas con algunas capas y relaciones, correcciones generales

+++ src/main/java/com/novabank/dto/MovimientoDTO.java
@@ -3,6 +3,7 @@ package com.novabank.dto;
 import com.novabank.model.TipoMovimiento;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,7 +11,7 @@ public class MovimientoDTO {
 
     private Long id;
     private TipoMovimiento tipo;
-    private Double cantidad;
+    private BigDecimal cantidad;
     private LocalDateTime fecha;
     private String numeroCuenta;
 }
